docker/requirements.txt
fastapi
joblib
pydantic
pyyaml
sklearn
uvicorn
